@@ -4,7 +4,7 @@ import android.Manifest
 import android.content.Context
 import android.os.Bundle
 import android.os.IBinder
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.text.TextUtils
 import android.view.MotionEvent
 import android.view.View
