@@ -1,10 +1,8 @@
 package www.ble.sixsix.base;
 
-import android.support.annotation.NonNull;
-
 public interface IHandler {
 
-    void handleData(@NonNull byte[] data);
+    void handleData(byte[] data);
 
     void reset();
 }
